complete file restructure
