+++ art/the_shop.rb
@@ -1,88 +1,6 @@
 # rubocop:disable all
 #-----------------------------YOUR CODE BELOW---------------------------------->
 
-def big_boss_art
-  art = { id: :move, art:
-  <<-'ASCII'
-                    / ___`\
-         /|        ( (   \ \
-    |^v^v  V|       \ \/) ) )
-    \  ____ /        \_/ / /
-    ,Y`    `,           / /
-    ||  -  -)          { }
-    \\   _\ |          | |
-     \\ / _`\_         / /
-     / |  ~ | ``\     _|_|
-  ,-`  \    |  \ \  ,//(_}
- /      |   |   | \/  \| |
-|     | \   /   | '   ,\ \
-\     |  | |    /\  _/` | |
- |    \  \ |   |  ``    | |
- |    |   |/   |        / /
-    ASCII
-  }
-end
-
-def the_barkeep
-  art = { id: :move, art: # this art is a combination of 2 pieces
-  <<-'ASCII'              # spacing is warped here for correct display on terminal
-                    _____
-                  __|___|__
-                   ('o_o')
-⢸⣷⣶⣤⡀              _\~-~/_    ______.   ⣤⣶⣾⡇
-⢸⡟⠛⠛⢻⣦⡀           //\__/\ \ ~(_]---' ⢀⣴⡟⠛⠛⢻⡇
-⢸⡟⠛⠛⠛⠛⢻⣶⣤⣀⡀      / )O  O( .\/_)  ⢀⣀⣤⣶⡟⠛⠛⠛⠛⢻⡇
-⢸⣟⣛⣛⣛⣛⣛⣛⣛⣛⣛⣻⡇    \ \    / \_/  ⢸⣟⣛⣛⣛⣛⣛⣛⣛⣛⣛⣛⣻⡇
-⢸⣏⣉⣉⣉⣉⣉⣉⣉⣉⣉⣻⡇    )/_|  |_\     ⢸⣟⣉⣉⣉⣉⣉⣉⣉⣉⣉⣉⣻⡇
-⢸⣏⣉⣉⣉⣉⣉⣉⣉⣉⣉⡏⢁⡀  // /(\/)\ \    ⢀⡈⢹⣉⣉⣉⣉⣉⣉⣉⣉⣉⣹⡇
-⢸⣏⣉⣉⣉⣉⣉⣉⣉⣉⣩⡇⢸⡇  /_/      \_\   ⢸⡇⢸⣍⣉⣉⣉⣉⣉⣉⣉⣉⣹⡇
- ⣯⣭⣭⣭⣭⣭⣭⣭⣭⣭⡇⠘⠃ (_||      ||_)  ⠘⠃⢸⣭⣭⣭⣭⣭⣭⣭⣭⣭⣻⡇
-⢸⣧⣤⣤⣤⣤⣤⣤⣤⣤⣬⣿⡇    \| |__| |/    ⠀⢸⣿⣥⣤⣤⣤⣤⣤⣤⣤⣤⣼⡇
-⢸⣧⣤⣤⣤⣤⣤⣤⣤⣤⣤⣿⡇     | |  | |      ⢸⣿⣤⣤⣤⣤⣤⣤⣤⣤⣼⡇
-⢸⣧⣤⣤⣤⣤⣤⣤⣤⣤⣤⣾⡇     |_|  |_|     ⠀⢸⣷⣤⣤⣤⣤⣤⣤⣤⣤⣷⡇
-                  /_\  /_\
-    ASCII
-     }
-end
-
-def the_bouncer
-  art = { id: :move, art:
-  <<-'ASCII'
-                     ______
-                   <((((((\\\
-                   /      . }\
-                   ;--..--._|}
-(\                 '--/\--'  )
- \\                | '-'  :'|
-  \\               . -==- .-|
-   \\               \.__.'   \--._
-   [\\          __.--|       //  _/'--.
-   \ \\       .'-._ ('-----'/ __/      \
-    \ \\     /   __>|      | '--.       |
-     \ \\   |   \   |     /    /       /
-      \ '\ /     \  |     |  _/       /
-       \  \       \ |     | /        /
-        \  \       \       /        /
-    ASCII
-   }
-end
-
-def the_band
-  art = { id: :move, art:
-    <<-'ASCII'
-
-    #####\             _             /#####
-    #( )# |          _( )__         | #( )#
-    ##### |         /_    /         | #####
-    #" "# |     ___m/I_ //_____     | #" "#
-    # O # |____#-x.\ /++m\ /.x-#____| # O #
-    #m.m# |   /" \ ///###\\\ / "\   | #m.m#
-   /#####/    ######/     \######    \#####\
-  /                                         \
-      ASCII
-  }
-end
-
 def the_shop
   art = [
     # ASCII art 1
