--- boss/messages/talks_with_the_boss.rb
@@ -1,69 +1,0 @@
-# rubocop:disable all
-#-----------------------------YOUR CODE BELOW---------------------------------->
-
-# end of main game rewards message
-
-def your_rewards(player, health, boost, multiplier)
-  boost = boost == :attack ? "💢 attack" : "🛡️ block"
-
-  messages = "#{HP_PLUS} +#{health} #{player[:emoji]}  #{BUFF} +#{multiplier} #{boost}"
-
-  puts text_break("🎉🪅 #{BONUS} 🎊🎈", " ", 80)
-  puts text_break(messages, " ", 90)
-end
-
-def greeting(place)
-  enemy = [
-    "Strike first, strike hard, no mercy",
-    "Step on up and show who's the boss",
-    "Will you get got, or you gone get?",
-    "Time for a thrashing",
-    "Lay down the law",
-  ]
-  shop = [
-    "🧞💬 山🝗㇄⼕ㄖ爪🝗 丂〸尺闩𝓝Ꮆ🝗尺...",
-  ]
-
-  x, messages = place == :combat ? [75, enemy] : [70, shop]
-  puts text_break(messages.sample, " ", x)
-end
-
-# Intro outro messages for changes to boss fighting style
-
-def boss_walks(the_boss, boss_style, time)
-  barkeep = [
-    "The #{boss_style} is keeping the liquor flowng, its happy hour ⌚",
-    "#{the_boss[:name]} is getting close to barring you, prepare!",
-    "The #{boss_style} is in the house 🛖 orders up",
-    "#{the_boss[:name]} is the liquor 🍾",
-  ]
-  bouncer = [ "#{the_boss[:name]} is excerting its pressure #{HANGOVER} +1 🍺" ]
-  band = [
-    "#{the_boss[:name]} is two-stepping 🪜 in the pit. Parkway Drive 🎸 come out on stage!",
-    "#{the_boss[:name]} is smashing out banging tunes 🎸 A pit is 'bout to break out!",
-    "Some wanker just started playing dubstep 🎧 Get ready for damage",
-    "Darude Sandstorm 🎧 starts playing, shit's 'bout to go down",
-  ]
-  outro = [ "#{the_boss[:name]} ♻️ ▻◅▶▷ #{boss_style}" ]
-  reprise = [ "#{HANGOVER} -1 🍺" ]
-
-  x, messages = case boss_style
-  when "🍻 Barkeep" then time == :intro ? [80, barkeep] : [85, outro]
-  when "🕶️ Bouncer" then time == :intro ? [100, bouncer] : [85, outro]
-  when "🎶 Band"    then (time == :reprise) ? [90, reprise] : [85, (time == :intro) ? band : outro]
-  end
-  puts text_break(messages.sample, " ", x)
-end
-
-def boss_speaks(player, the_boss, blast = 0, outcome)
-  hit = [
-    "#{the_boss[:name]} has had enough of your crap 🔥 #{RAGE} 💢 #{player[:name]} -#{blast} #{player[:emoji]}",
-  ]
-  threat = [
-    "#{the_boss[:name]} notices your cowardly attempts, a thrashing #{RAGE} is on the cards",
-    "#{the_boss[:name]} is getting aggy #{RAGE} You best tread lightly",
-  ]
-
-  messages = outcome == :hit ? hit : threat
-  puts text_break(messages.sample, " ", 100)
-end
