+++ combat/beatdown.rb
