+++ combat/boss_band.rb
