+++ combat/boss_barkeep.rb
