+++ combat/boss_bouncer.rb
