+++ messages/war_chatter.rb
@@ -59,3 +59,51 @@ def enemy_speaks(enemy, event)
   puts text_break(messages, " ", x)
 
 end
+
+# Boss Rage
+
+def boss_speaks(player, the_boss, blast = 0, outcome)
+  hit = [
+    "#{the_boss[:name]} has had enough of your crap 🔥 #{RAGE} 💢 #{player[:name]} -#{blast} #{player[:emoji]}",
+  ]
+  threat = [
+    "#{the_boss[:name]} notices your cowardly attempts, a thrashing #{RAGE} is on the cards",
+    "#{the_boss[:name]} is getting aggy #{RAGE} You best tread lightly",
+  ]
+
+  messages = outcome == :hit ? hit : threat
+  puts text_break(messages.sample, " ", 100)
+end
+
+# Weapon Messages
+
+def weapon_speaks(weapon, status)
+  got = [
+    " There's enough useless crap 🛒 here to make a #{weapon[:name]}, your mad smithing skills allow it",
+    " Father Christmas 🎅 brings your gift early, it's that #{weapon[:name]} you wished for",
+    " Gift! There's a #{weapon[:name]} here! Looks like the foot's on the other shoe!",
+    " You reach into your pocket for a snout 🚬 and pull out a #{weapon[:name]}!",
+    " You remembered you still have your #{weapon[:name]} in your sock 🧦",
+    " 🕵🏼 Mysterious Stranger 💬 Hey kid, catch! #{weapon[:name]} acquired",
+    " Package from Amazon 🎁 It's that #{weapon[:name]} you ordered!",
+    " There's find a #{weapon[:name]} sandwiched between the pickles",
+    " There's a #{weapon[:name]} inside this hollowed out book 📙",
+    " A Djinn 🧞 grants you a wish, one #{weapon[:name]} please",
+    " You found a #{weapon[:name]} hidden among the spoons!",
+  ]
+  broke = [
+    " The universe conspires against you. Your #{weapon[:name]} melts!",
+    " Your #{weapon[:name]} jumps ouuta your hand and runs off",
+    " You suddenly forgot how to wield your #{weapon[:name]}",
+    " Oh shit! Your #{weapon[:name]} broke. Better leg it!",
+    " Poxy #{weapon[:name]} gon 'n' did a runner on you",
+    " Your #{weapon[:name]} disappears into thin air",
+    " Your #{weapon[:name]} flew out of the window!",
+    " Goodbye #{weapon[:name]}, you will be missed",
+    " Your #{weapon[:name]} resigned from service",
+    " Damn it! You dropped your #{weapon[:name]}!",
+    " Your #{weapon[:name]} was too shoddy",
+  ]
+  messages = status == :got ? (WEAPON + got.sample) : (BROKE + broke.sample)
+  puts text_break(messages, " ", 100)
+end
