+++ powers/blackjack.rb
